requirements.txt
streamlit
mediapipe
opencv-python-headless
numpy
langchain-groq
langgraph
Pillow